Extract PersistentWallet to a file
